Free memory in parent & child before exiting
By default, conmand runs as a daemon process in the background.
This is performed via a traditional double-fork technique to detach
from the terminal.  Since the parent and child processes quickly
exit during this daemonization, memory cleanup is left for the
runtime system.  This causes valgrind to report the following leaks:

> HEAP SUMMARY:
>     in use at exit: 520 bytes in 6 blocks
>   total heap usage: 11 allocs, 5 frees, 699 bytes allocated
>
> 23 bytes in 1 blocks are still reachable in loss record 1 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x5B05989: strdup (strdup.c:42)
>    by 0x4071EC: parse_cmdline (conf.c:311)
>    by 0x403AB2: main (munged.c:115)
>
> 82 bytes in 1 blocks are still reachable in loss record 2 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x5B05989: strdup (strdup.c:42)
>    by 0x4073F0: parse_cmdline (conf.c:361)
>    by 0x403AB2: main (munged.c:115)
>
> 82 bytes in 1 blocks are still reachable in loss record 3 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x5B05989: strdup (strdup.c:42)
>    by 0x407457: parse_cmdline (conf.c:368)
>    by 0x403AB2: main (munged.c:115)
>
> 82 bytes in 1 blocks are still reachable in loss record 4 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x5B05989: strdup (strdup.c:42)
>    by 0x407624: parse_cmdline (conf.c:396)
>    by 0x403AB2: main (munged.c:115)
>
> 83 bytes in 1 blocks are still reachable in loss record 5 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x5B05989: strdup (strdup.c:42)
>    by 0x40768B: parse_cmdline (conf.c:403)
>    by 0x403AB2: main (munged.c:115)
>
> 168 bytes in 1 blocks are still reachable in loss record 6 of 6
>    at 0x4C28C20: malloc (vg_replace_malloc.c:296)
>    by 0x4069FF: create_conf (conf.c:133)
>    by 0x403A93: main (munged.c:114)
>
> LEAK SUMMARY:
>    definitely lost: 0 bytes in 0 blocks
>    indirectly lost: 0 bytes in 0 blocks
>      possibly lost: 0 bytes in 0 blocks
>    still reachable: 520 bytes in 6 blocks
>         suppressed: 0 bytes in 0 blocks
>
> For counts of detected and suppressed errors, rerun with: -v
> ERROR SUMMARY: 6 errors from 6 contexts (suppressed: 0 from 0)

Now, destroy_conf() is called during daemonization immediately before
the parent and child processes exit, thereby freeing memory used by
the configuration.  Special care is taken to avoid unlinking the PID
file during these early calls to destroy_conf() since this file must
remain until the grandchild process exits when the daemon is shut down.

Tested on:
- CentOS 7.4.1708
- Debian 8.10, 9.3
- Fedora 27
- Ubuntu 16.04.3 LTS, 17.10

+++ src/munged/conf.c
@@ -203,7 +203,7 @@ create_conf (void)
 
 
 void
-destroy_conf (conf_t conf)
+destroy_conf (conf_t conf, int do_unlink)
 {
     assert (conf != NULL);
     assert (conf->ld < 0);              /* sock_destroy() already called */
@@ -222,7 +222,9 @@ destroy_conf (conf_t conf)
         conf->logfile_name = NULL;
     }
     if (conf->pidfile_name) {
-        (void) unlink (conf->pidfile_name);
+        if (do_unlink) {
+            (void) unlink (conf->pidfile_name);
+        }
         free (conf->pidfile_name);
         conf->pidfile_name = NULL;
     }
